+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/LuasLingkaran.kt
@@ -1,5 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class LuasLingkaran(
-    val hasil: Long
+    val hasil: Float
 )
