+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/LuasPersegi.kt
@@ -1,8 +1,7 @@
 package com.mopro.rumusdasarmatermatika.model
 
-import java.math.BigInteger
 
 data class LuasPersegi(
-    val hasil: Long
+    val hasil: Float
 
 )
