+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/LuasPersegiPanjang.kt
@@ -1,5 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class LuasPersegiPanjang(
-    val hasil: Long
+    val hasil: Float
 )
